+++ opendbc/sunnypilot/car/hyundai/longitudinal/controller.py
@@ -11,11 +11,16 @@ from dataclasses import dataclass
 from opendbc.car import structs, DT_CTRL
 from opendbc.car.interfaces import CarStateBase
 from opendbc.car.hyundai.values import CarControllerParams
-from opendbc.sunnypilot.car.hyundai.longitudinal.helpers import jerk_limited_integrator
+from opendbc.sunnypilot.car.hyundai.longitudinal.helpers import get_car_config, jerk_limited_integrator, ramp_update
 from opendbc.sunnypilot.car.hyundai.longitudinal.tuning_controller import LongitudinalTuningController
 from opendbc.sunnypilot.car.hyundai.values import HyundaiFlagsSP
 
 LongCtrlState = structs.CarControl.Actuators.LongControlState
+
+MIN_JERK = 0.5
+
+GEN1_LOWER_JERK_BP = [-2.0, -1.5, -1.0, -0.25, -0.1, -0.025, -0.01, -0.005]
+GEN1_LOWER_JERK_V  = [ 3.3,  2.5,  2.0,   1.9,  1.8,   1.65,  1.15,    0.5]
 
 
 @dataclass
@@ -35,7 +40,7 @@ class LongitudinalController:
     self.CP = CP
     self.CP_SP = CP_SP
 
-    self.tuning = LongitudinalTuningController(CP, CP_SP)
+    self.car_config = get_car_config(CP)
     self.long_control_state_last = LongCtrlState.off
     self.stopping_count = 0
 
@@ -71,16 +76,64 @@ class LongitudinalController:
     self.stopping_count += 1
 
   def calculate_jerk(self, CC: structs.CarControl, CS: CarStateBase, long_control_state: LongCtrlState) -> None:
-    """Calculate jerk based on tuning."""
-    self.tuning.calculate_jerk(CC, CS, long_control_state)
+    if not self.CP_SP.flags & HyundaiFlagsSP.LONG_TUNING:
+      jerk_limit = 3.0 if long_control_state == LongCtrlState.pid else 1.0
+
+      self.jerk_upper = jerk_limit
+      self.jerk_lower = 5.0
+      return
+
+    velocity = CS.out.vEgo
+    accel_error = self.accel_cmd - self.accel_last
+
+    a_ego_blended = float(np.interp(velocity, [1.0, 2.0], [CS.aBasis, CS.out.aEgo]))
+
+    # Lookahead jerk: How much jerk is needed to reach desired accel in future_t seconds
+    future_t_upper = float(np.interp(velocity, self.car_config.lookahead_jerk_bp, self.car_config.lookahead_jerk_upper_v))
+    future_t_lower = float(np.interp(velocity, self.car_config.lookahead_jerk_bp, self.car_config.lookahead_jerk_lower_v))
+
+    # Required jerk to reach target accel in lookahead window
+    j_ego_upper = accel_error / future_t_upper
+    j_ego_lower = accel_error / future_t_lower
+
+    # Jerk is limited by the following conditions imposed by ISO 15622:2018.
+    upper_speed_factor = 0.5
+    if long_control_state == LongCtrlState.pid:
+      upper_speed_factor = float(np.interp(velocity, [0.0, 5.0, 20.0], [2.0, 3.0, 1.6]))
+    lower_speed_factor = float(np.interp(velocity, [0.0, 5.0, 20.0], [5.0, 5.0, 2.5]))
+
+    lower_jerk = max(-j_ego_lower, MIN_JERK)
+    if self.CP.radarUnavailable:
+      lower_jerk = 5.0
+
+    # Final jerk limits with thresholds
+    desired_jerk_upper = min(max(j_ego_upper, MIN_JERK), upper_speed_factor)
+    desired_jerk_lower = min(lower_jerk, lower_speed_factor)
+
+    # Gen1 negative accel tune
+    gen1_accel_error = a_ego_blended - self.accel_last
+    if self.CP.radarUnavailable:
+      gen1_lower_jerk = 5.0
+    elif gen1_accel_error < 0:
+      lower_max = self.car_config.jerk_limits
+      original_value = np.array(GEN1_LOWER_JERK_V)
+      dynamic_lower = original_value * (lower_max / original_value[0])
+      gen1_lower_jerk = float(np.interp(gen1_accel_error, GEN1_LOWER_JERK_BP, dynamic_lower))
+    else:
+      gen1_lower_jerk = 0.5
+    gen1_desired_jerk_lower = min(gen1_lower_jerk, lower_speed_factor)
+
+    # New gen tune uses LONG_TUNING_BRAKING flag
+    if self.CP_SP.flags & HyundaiFlagsSP.LONG_TUNING_BRAKING:
+      self.jerk_upper = ramp_update(self.jerk_upper, desired_jerk_upper)
+      self.jerk_lower = desired_jerk_lower
+    else:
+      self.jerk_upper = ramp_update(self.jerk_upper, desired_jerk_upper)
+      self.jerk_lower = ramp_update(self.jerk_lower, gen1_desired_jerk_lower)
 
     if not CC.longActive:
       self.jerk_upper = 0.0
       self.jerk_lower = 0.0
-      return
-
-    self.jerk_upper = self.tuning.jerk_upper
-    self.jerk_lower = self.tuning.jerk_lower
 
   def calculate_a_value(self, CC: structs.CarControl) -> None:
     if not self.CP_SP.flags & HyundaiFlagsSP.LONG_TUNING or self.CP.radarUnavailable:
