+++ opendbc/sunnypilot/car/hyundai/longitudinal/helpers.py
@@ -5,9 +5,14 @@ This file is part of sunnypilot and is licensed under the MIT License.
 See the LICENSE.md file in the root directory for more details.
 """
 
+import numpy as np
+
 from opendbc.car import structs, DT_CTRL, rate_limit
 from opendbc.car.hyundai.values import HyundaiFlags
 from opendbc.sunnypilot.car.hyundai.longitudinal.longitudinal_config import CarTuningConfig, TUNING_CONFIGS, CAR_SPECIFIC_CONFIGS
+
+JERK_THRESHOLD = 0.1
+JERK_STEP = 0.1
 
 
 def get_car_config(CP: structs.CarParams) -> CarTuningConfig:
@@ -43,3 +48,10 @@ def jerk_limited_integrator(desired_accel, last_accel, jerk_upper, jerk_lower) -
     val = jerk_lower * DT_CTRL * 2
 
   return rate_limit(desired_accel, last_accel, -val, val)
+
+
+def ramp_update(current, target):
+  error = target - current
+  if abs(error) > JERK_THRESHOLD:
+    return current + float(np.clip(error, -JERK_STEP, JERK_STEP))
+  return target
